+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## WIP mobile_dropdown
+
+### Added
+
+- Add `animate-out` option to Drop component
+- Add alignment `stretch` to Drop component
+- Add Height Viewport uses scroll parent as viewport 
+- Add zero z-index utility class to Position component
+
+### Fixed
+
+- Make content clickable for sticky cover and reveal effects
+
 ## WIP
 
 ### Added